Sysinfo util
This commit moves sysinfo collectibles to the avocado utils. It creates
api for collecting sysinfo. This api will be used in sysinfo runner and
sysinfo plugin. This is part of work on sysinfo support for nrunner.

Reference: #3877
